General system information

The get_sys_info returns general system level information as a list. The function parameters control the information to upload. If a parameter is set to FALSE, an NA is uploaded instead. Each element of the list is contains the output from:

Sys.info();

get_platform_info();

get_r_version();

get_ram();

get_cpu();

get_byte_compiler();

get_linear_algebra();

Sys.getlocale()

installed.packages();

.Machine

The package version number;

Unique ID - used to extract results;

The current date.

get_sys_details(
  sys_info = TRUE,
  platform_info = TRUE,
  r_version = TRUE,
  ram = TRUE,
  cpu = TRUE,
  byte_compiler = TRUE,
  linear_algebra = TRUE,
  locale = TRUE,
  installed_packages = TRUE,
  machine = TRUE
)

Arguments

sys_info

Default TRUE.

platform_info

Default TRUE.

r_version

Default TRUE.

ram

Default TRUE.

cpu

Default TRUE.

byte_compiler

Default TRUE.

linear_algebra

Default TRUE.

locale

Default TRUE

installed_packages

Default TRUE.

machine

Default TRUE

Value

A list

Examples

## Returns all details about your machine
get_sys_details(cpu = FALSE, installed_packages = FALSE, ram = FALSE)
#> $sys_info
#> $sys_info$sysname
#> [1] "Darwin"
#> 
#> $sys_info$release
#> [1] "20.6.0"
#> 
#> $sys_info$version
#> [1] "Darwin Kernel Version 20.6.0: Tue Apr 19 21:04:45 PDT 2022; root:xnu-7195.141.29~1/RELEASE_X86_64"
#> 
#> $sys_info$nodename
#> [1] "Mac-1655283188675.local"
#> 
#> $sys_info$machine
#> [1] "x86_64"
#> 
#> $sys_info$login
#> [1] "root"
#> 
#> $sys_info$user
#> [1] "runner"
#> 
#> $sys_info$effective_user
#> [1] "runner"
#> 
#> 
#> $platform_info
#> $platform_info$OS.type
#> [1] "unix"
#> 
#> $platform_info$file.sep
#> [1] "/"
#> 
#> $platform_info$dynlib.ext
#> [1] ".so"
#> 
#> $platform_info$GUI
#> [1] "X11"
#> 
#> $platform_info$endian
#> [1] "little"
#> 
#> $platform_info$pkgType
#> [1] "mac.binary"
#> 
#> $platform_info$path.sep
#> [1] ":"
#> 
#> $platform_info$r_arch
#> [1] ""
#> 
#> 
#> $r_version
#> $r_version$platform
#> [1] "x86_64-apple-darwin17.0"
#> 
#> $r_version$arch
#> [1] "x86_64"
#> 
#> $r_version$os
#> [1] "darwin17.0"
#> 
#> $r_version$system
#> [1] "x86_64, darwin17.0"
#> 
#> $r_version$status
#> [1] ""
#> 
#> $r_version$major
#> [1] "4"
#> 
#> $r_version$minor
#> [1] "2.0"
#> 
#> $r_version$year
#> [1] "2022"
#> 
#> $r_version$month
#> [1] "04"
#> 
#> $r_version$day
#> [1] "22"
#> 
#> $r_version$`svn rev`
#> [1] "82229"
#> 
#> $r_version$language
#> [1] "R"
#> 
#> $r_version$version.string
#> [1] "R version 4.2.0 (2022-04-22)"
#> 
#> $r_version$nickname
#> [1] "Vigorous Calisthenics"
#> 
#> 
#> $ram
#> [1] NA
#> 
#> $cpu
#> [1] NA
#> 
#> $byte_compiler
#> byte_optimize 
#>             2 
#> 
#> $linear_algebra
#> $linear_algebra$blas
#> [1] "/Library/Frameworks/R.framework/Versions/4.2/Resources/lib/libRblas.0.dylib"
#> 
#> $linear_algebra$lapack
#> [1] "/Library/Frameworks/R.framework/Versions/4.2/Resources/lib/libRlapack.dylib"
#> 
#> 
#> $locale
#> [1] "C/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8"
#> 
#> $installed_packages
#> [1] NA
#> 
#> $machine
#> $machine$double.eps
#> [1] 2.220446e-16
#> 
#> $machine$double.neg.eps
#> [1] 1.110223e-16
#> 
#> $machine$double.xmin
#> [1] 2.225074e-308
#> 
#> $machine$double.xmax
#> [1] 1.797693e+308
#> 
#> $machine$double.base
#> [1] 2
#> 
#> $machine$double.digits
#> [1] 53
#> 
#> $machine$double.rounding
#> [1] 5
#> 
#> $machine$double.guard
#> [1] 0
#> 
#> $machine$double.ulp.digits
#> [1] -52
#> 
#> $machine$double.neg.ulp.digits
#> [1] -53
#> 
#> $machine$double.exponent
#> [1] 11
#> 
#> $machine$double.min.exp
#> [1] -1022
#> 
#> $machine$double.max.exp
#> [1] 1024
#> 
#> $machine$integer.max
#> [1] 2147483647
#> 
#> $machine$sizeof.long
#> [1] 8
#> 
#> $machine$sizeof.longlong
#> [1] 8
#> 
#> $machine$sizeof.longdouble
#> [1] 16
#> 
#> $machine$sizeof.pointer
#> [1] 8
#> 
#> $machine$longdouble.eps
#> [1] 1.084202e-19
#> 
#> $machine$longdouble.neg.eps
#> [1] 5.421011e-20
#> 
#> $machine$longdouble.digits
#> [1] 64
#> 
#> $machine$longdouble.rounding
#> [1] 5
#> 
#> $machine$longdouble.guard
#> [1] 0
#> 
#> $machine$longdouble.ulp.digits
#> [1] -63
#> 
#> $machine$longdouble.neg.ulp.digits
#> [1] -64
#> 
#> $machine$longdouble.exponent
#> [1] 15
#> 
#> $machine$longdouble.min.exp
#> [1] -16382
#> 
#> $machine$longdouble.max.exp
#> [1] 16384
#> 
#> 
#> $package_version
#> [1] "1.0.8"
#> 
#> $id
#> [1] "2022-06-15-76050539"
#> 
#> $date
#> [1] "2022-06-15"
#>