Upload benchmark results

This function uploads the benchmarking results. These results will then be incorparated in future versions of the package.

create_bundle(results, filename = NULL, args = NULL, id_prefix = "")

upload_results(
  results,
  url = "http://www.mas.ncl.ac.uk/~ncsg3/form.php",
  args = NULL,
  id_prefix = ""
)

Arguments

results

Benchmark results. Probably obtained from benchmark_std() or benchmark_io().

filename

default NULL. A character vector of where to store the results (in an .rds file). If NULL, results are not saved.

args

Default NULL. A list of arguments to be passed to get_sys_details().

id_prefix

Character string to prefix the benchmark id. Makes it easier to retrieve past results.

url

The location of where to upload the results.

Examples

## Run benchmarks
if (FALSE) {
res = benchmark_std()
upload_results(res)
}